more warnings and options

+++ make-wiki-search.sh
@@ -1,18 +1,28 @@
 #!/bin/bash
 
-RP="rampart"
-
-EXTRACTOR="./WikiExtractor.py"
-
-DATADIR="./wikidata"
-DBDIR="./wikidb"
-HAVEPV=""
-
 die () {
 	echo $1
 	exit 1
 }
 
+# Set the name of the user that web server will run under.
+# This is the name of the account that will start the webserver
+#   or the name set in web_server_conf.js if started as root.
+
+WEBUSER="nobody"
+
+RP=`which rampart`;
+
+if [ "$RP" == "" ]; then
+    die "Can't find rampart executable"
+fi
+
+ME=`whoami`
+
+EXTRACTOR="./WikiExtractor.py"
+
+DATADIR="./wikidata"
+HAVEPV=""
 
 curl --version &>/dev/null || die "curl must be installed and in the current \$PATH before running this script"
 
@@ -41,12 +51,25 @@ curl -C - -o enwiki-latest-pages-articles.xml.bz2 https://dumps.wikimedia.org/en
 
 echo "Decompressing..."
 
-if [ -e enwiki-latest-pages-articles.xml ]; then
+check_reply () {
     echo "enwiki-latest-pages-articles.xml exists.";
     read -p "[o]verwrite or [c]ontinue with existing" -n 1 -r
+    echo
+}
+
+if [ -e enwiki-latest-pages-articles.xml ]; then
+    REPLY="";
+    while [[ ! $REPLY  =~ ^[oO]$ ]] && [[ ! $REPLY  =~ ^[cC]$ ]]; do
+       if [ "$REPLY" != "" ]; then
+            echo "invalid response"
+       fi
+       check_reply
+    done
+
     if [[ $REPLY =~ ^[oO]$ ]]; then
 	echo
         if [ "$HAVEPV" == "1" ] ; then
+                # we need to cat the file since it is too large for bzcat on a 32bit system.
                 cat enwiki-latest-pages-articles.xml.bz2 | pv -s $(ls -l enwiki-latest-pages-articles.xml.bz2 | awk '{print $5}') | bzcat -d > enwiki-latest-pages-articles.xml || die "Failed to decompress file"
         else
                 cat enwiki-latest-pages-articles.xml.bz2 | bzcat -d > enwiki-latest-pages-articles.xml || die "Failed to decompress file"
@@ -60,18 +83,43 @@ else
     fi
 fi
 
-echo "Extracting text from enwiki-latest-pages-articles.xml"
+check_datadir () {
+    if [ -e $DATADIR/txt/AA ] ; then 
+        echo "There appears to be text already extracted in the $DATADIR/txt directory"
+        echo "Should we overwrite that data and start again or use the data already present?"
+        read -p "[o]verwrite or [c]ontinue with existing" -n 1 -r
+        echo
+    fi
+}
 
-mkdir -p "$DATADIR/txt" || die "could not make directory $DATADIR/txt"
+REPLY="";
+while [[ ! $REPLY  =~ ^[oO]$ ]] && [[ ! $REPLY  =~ ^[cC]$ ]]; do
+   if [ "$REPLY" != "" ]; then
+        echo "invalid response"
+   fi
+   check_datadir
+done
 
-#./WikiExtractor.py -o "$DATADIR/txt" enwiki-latest-pages-articles.xml|| die "failed to extract text from enwiki-latest-pages-articles.xml"
-./WikiExtractor.py -o "$DATADIR/txt" enwiki-latest-pages-articles.xml 2>&1 | tee extractor-output.txt | while read i; do 
-	line=$(echo -n $i | grep -oE '[[:digit:]]+.+'); 
-	printf "%s            \r" "$line"; 
-done || die "failed to extract text from enwiki-latest-pages-articles.xml"
+if [[ $REPLY =~ ^[oO]$ ]]; then
+    echo "removing content in $DATADIR/txt/" 
+    rm -rf $DATADIR/txt/*
+    REPLY="";
+fi
 
-if [ ! -e ./web_server/data/docs ]; then
-    mkdir -p ./web_server/data/docs || die "could not create directory ./web_server/data/docs"
+if [[ $REPLY =~ ^[Cc]$ ]]; then
+    echo "Extracting text from enwiki-latest-pages-articles.xml"
+
+    mkdir -p "$DATADIR/txt" || die "could not make directory $DATADIR/txt"
+
+    #./WikiExtractor.py -o "$DATADIR/txt" enwiki-latest-pages-articles.xml|| die "failed to extract text from enwiki-latest-pages-articles.xml"
+    ./WikiExtractor.py -o "$DATADIR/txt" enwiki-latest-pages-articles.xml 2>&1 | tee extractor-output.txt | while read i; do 
+        line=$(echo -n $i | grep -oE '[[:digit:]]+.+'); 
+        printf "%s            \r" "$line"; 
+    done || die "failed to extract text from enwiki-latest-pages-articles.xml"
+fi
+
+if [ ! -e ./web_server/data ]; then
+    mkdir -p ./web_server/data || die "could not create directory ./web_server/data"
 fi
 
 
@@ -79,6 +127,16 @@ echo "importing data"
 $RP import.js
 echo "creating text index"
 $RP mkindex.js
+
+if [ "$ME" == "root" ]; then
+    chown -R nobody ./web_server/data
+fi
+
+if [ "$ME" != "$WEBUSER" ] ;
+    echo "WARNING: the database was created as user '$ME', but the webserver is set to be";
+    echo "         run as '$WEBUSER'.  Please check and correct the owner of the "
+    echo "         './web_server/data/wikipedia_search/' directory and files therein."
+
 
 #echo "Now you can run \"rampart import.js\" and \"rampart mkindex.js\" to create the database."
 #echo "After it is made, \"rampart wksearch.js\" will start the server"
